refactor: Move GFDMOperator to _compat module (#844)
Old import path preserved with deprecation warning. 51 tests pass.
