+++ mfgarchon/utils/numerical/_compat/__init__.py
@@ -1,0 +1,6 @@
+"""
+Backward-compatible imports for deprecated GFDM modules.
+
+GFDMOperator has been superseded by TaylorOperator from gfdm_strategies.py.
+This module preserves the import path for existing code.
+"""
